LibJS: Move ExecutionContext function implementations out of line
