+++ Userland/Libraries/LibJS/Runtime/ExecutionContext.h
@@ -1,6 +1,7 @@
 /*
  * Copyright (c) 2020-2021, Andreas Kling <kling@serenityos.org>
  * Copyright (c) 2020-2021, Linus Groh <linusg@serenityos.org>
+ * Copyright (c) 2022, Luke Wilde <lukew@serenityos.org>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
@@ -21,34 +22,12 @@ using ScriptOrModule = Variant<Empty, NonnullGCPtr<Script>, NonnullGCPtr<Module>
 
 // 9.4 Execution Contexts, https://tc39.es/ecma262/#sec-execution-contexts
 struct ExecutionContext {
-    explicit ExecutionContext(Heap& heap)
-        : arguments(heap)
-    {
-    }
+    explicit ExecutionContext(Heap& heap);
 
-    [[nodiscard]] ExecutionContext copy() const
-    {
-        ExecutionContext copy { arguments };
-
-        copy.function = function;
-        copy.realm = realm;
-        copy.script_or_module = script_or_module;
-        copy.lexical_environment = lexical_environment;
-        copy.variable_environment = variable_environment;
-        copy.private_environment = private_environment;
-        copy.current_node = current_node;
-        copy.function_name = function_name;
-        copy.this_value = this_value;
-        copy.is_strict_mode = is_strict_mode;
-
-        return copy;
-    }
+    [[nodiscard]] ExecutionContext copy() const;
 
 private:
-    explicit ExecutionContext(MarkedVector<Value> existing_arguments)
-        : arguments(move(existing_arguments))
-    {
-    }
+    explicit ExecutionContext(MarkedVector<Value> existing_arguments);
 
 public:
     FunctionObject* function { nullptr };                // [[Function]]
